Change the Java equivalent of the Number type to float.

+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/AutoCompleteTag.java
@@ -62,11 +62,11 @@ public class AutoCompleteTag extends WidgetTag /* interfaces */implements Animat
         setProperty("dataTextField", value);
     }
 
-    public int getDelay() {
-        return (int)getProperty("delay");
+    public float getDelay() {
+        return (float)getProperty("delay");
     }
 
-    public void setDelay(int value) {
+    public void setDelay(float value) {
         setProperty("delay", value);
     }
 
@@ -86,11 +86,11 @@ public class AutoCompleteTag extends WidgetTag /* interfaces */implements Animat
         setProperty("filter", value);
     }
 
-    public int getHeight() {
-        return (int)getProperty("height");
+    public float getHeight() {
+        return (float)getProperty("height");
     }
 
-    public void setHeight(int value) {
+    public void setHeight(float value) {
         setProperty("height", value);
     }
 
@@ -110,11 +110,11 @@ public class AutoCompleteTag extends WidgetTag /* interfaces */implements Animat
         setProperty("ignoreCase", value);
     }
 
-    public int getMinLength() {
-        return (int)getProperty("minLength");
+    public float getMinLength() {
+        return (float)getProperty("minLength");
     }
 
-    public void setMinLength(int value) {
+    public void setMinLength(float value) {
         setProperty("minLength", value);
     }
