+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/ComboBoxTag.java
@@ -86,11 +86,11 @@ public class ComboBoxTag extends WidgetTag /* interfaces */implements Animation,
         setProperty("dataValueField", value);
     }
 
-    public int getDelay() {
-        return (int)getProperty("delay");
+    public float getDelay() {
+        return (float)getProperty("delay");
     }
 
-    public void setDelay(int value) {
+    public void setDelay(float value) {
         setProperty("delay", value);
     }
 
@@ -110,11 +110,11 @@ public class ComboBoxTag extends WidgetTag /* interfaces */implements Animation,
         setProperty("filter", value);
     }
 
-    public int getHeight() {
-        return (int)getProperty("height");
+    public float getHeight() {
+        return (float)getProperty("height");
     }
 
-    public void setHeight(int value) {
+    public void setHeight(float value) {
         setProperty("height", value);
     }
 
@@ -134,19 +134,19 @@ public class ComboBoxTag extends WidgetTag /* interfaces */implements Animation,
         setProperty("ignoreCase", value);
     }
 
-    public int getIndex() {
-        return (int)getProperty("index");
+    public float getIndex() {
+        return (float)getProperty("index");
     }
 
-    public void setIndex(int value) {
+    public void setIndex(float value) {
         setProperty("index", value);
     }
 
-    public int getMinLength() {
-        return (int)getProperty("minLength");
+    public float getMinLength() {
+        return (float)getProperty("minLength");
     }
 
-    public void setMinLength(int value) {
+    public void setMinLength(float value) {
         setProperty("minLength", value);
     }
