+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/AreaTag.java
@@ -53,35 +53,35 @@ public class AreaTag extends BaseTag /* interfaces */implements Border/* interfa
         setProperty("background", value);
     }
 
-    public int getOpacity() {
-        return (int)getProperty("opacity");
+    public float getOpacity() {
+        return (float)getProperty("opacity");
     }
 
-    public void setOpacity(int value) {
+    public void setOpacity(float value) {
         setProperty("opacity", value);
     }
 
-    public int getHeight() {
-        return (int)getProperty("height");
+    public float getHeight() {
+        return (float)getProperty("height");
     }
 
-    public void setHeight(int value) {
+    public void setHeight(float value) {
         setProperty("height", value);
     }
 
-    public int getMargin() {
-        return (int)getProperty("margin");
+    public float getMargin() {
+        return (float)getProperty("margin");
     }
 
-    public void setMargin(int value) {
+    public void setMargin(float value) {
         setProperty("margin", value);
     }
 
-    public int getWidth() {
-        return (int)getProperty("width");
+    public float getWidth() {
+        return (float)getProperty("width");
     }
 
-    public void setWidth(int value) {
+    public void setWidth(float value) {
         setProperty("width", value);
     }
