+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/MarkersTag.java
@@ -53,11 +53,11 @@ public class MarkersTag extends BaseTag /* interfaces */implements Border/* inte
         setProperty("background", value);
     }
 
-    public int getSize() {
-        return (int)getProperty("size");
+    public float getSize() {
+        return (float)getProperty("size");
     }
 
-    public void setSize(int value) {
+    public void setSize(float value) {
         setProperty("size", value);
     }
