+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/TitleTag.java
@@ -69,19 +69,19 @@ public class TitleTag extends BaseTag /* interfaces */implements Border/* interf
         setProperty("font", value);
     }
 
-    public int getMargin() {
-        return (int)getProperty("margin");
+    public float getMargin() {
+        return (float)getProperty("margin");
     }
 
-    public void setMargin(int value) {
+    public void setMargin(float value) {
         setProperty("margin", value);
     }
 
-    public int getPadding() {
-        return (int)getProperty("padding");
+    public float getPadding() {
+        return (float)getProperty("padding");
     }
 
-    public void setPadding(int value) {
+    public void setPadding(float value) {
         setProperty("padding", value);
     }
 
@@ -93,11 +93,11 @@ public class TitleTag extends BaseTag /* interfaces */implements Border/* interf
         setProperty("position", value);
     }
 
-    public int getRotation() {
-        return (int)getProperty("rotation");
+    public float getRotation() {
+        return (float)getProperty("rotation");
     }
 
-    public void setRotation(int value) {
+    public void setRotation(float value) {
         setProperty("rotation", value);
     }
